backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.42.1
google-auth==2.26.1
google-auth-oauthlib==1.2.0
PyJWT==2.8.1
python-multipart==0.0.6
aiohttp==3.9.1
requests==2.31.0
